backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
openai==1.60.2 